finshing the admin and user auth test

+++ routes/user/usersList.js
@@ -4,9 +4,10 @@ import { ensureLoggedIn } from 'connect-ensure-login'
 const router = Router()
 export function usersList() {
   router.get('/users', ensureLoggedIn('/'), async (req, res, next) => {
-    if (req.session.user.occupation != 'admin') {
-      return res.redirect('/')
+    if (req.session.user.occupation !== 'admin') {
+      return res.status(302).redirect('/')
     }
+
     try {
       const users = await UserController.getUsers()
       const userList = await Promise.all(
@@ -19,7 +20,7 @@ export function usersList() {
           return userData
         }),
       )
-      return res.render('user/users', {
+      return res.status(200).render('user/users', {
         page: 'users',
         users: userList,
       })
@@ -27,7 +28,11 @@ export function usersList() {
       return next(error)
     }
   })
-  router.get('/delete/:id', async (req, res, next) => {
+  router.get('/delete/:id', ensureLoggedIn('/'), async (req, res, next) => {
+    // console.log('u', req.session)
+    if (req.session.user.occupation !== 'admin') {
+      return res.status(302).redirect('/')
+    }
     try {
       await UserController.deleteUser(req.params.id)
       req.session.messages.push({
@@ -39,5 +44,24 @@ export function usersList() {
       return next(error)
     }
   })
+  router.delete(
+    '/delete/:username',
+    ensureLoggedIn('/'),
+    async (req, res, next) => {
+      if (req.session.user.occupation !== 'admin') {
+        return res.status(302).redirect('/')
+      }
+      try {
+        await UserController.deleteUserByUsername(req.params.username)
+        req.session.messages.push({
+          text: 'The user was deleted successfully',
+          type: 'info',
+        })
+        return res.redirect('/user/users')
+      } catch (error) {
+        return next(error)
+      }
+    },
+  )
   return router
 }
